+++ AudioKit/Common/Nodes/Generators/Oscillators/Oscillator Bank/AKOscillatorBank.swift
@@ -202,11 +202,11 @@ open class AKOscillatorBank: AKPolyphonicNode, AKComponent {
 
     /// Function to start, play, or activate the node, all do the same thing
     open override func play(noteNumber: MIDINoteNumber, velocity: MIDIVelocity) {
-        self.internalAU!.startNote(Int32(noteNumber), velocity: Int32(velocity))
+        self.internalAU!.startNote(noteNumber, velocity: velocity)
     }
 
     /// Function to stop or bypass the node, both are equivalent
     open override func stop(noteNumber: MIDINoteNumber) {
-        self.internalAU!.stopNote(Int32(noteNumber))
+        self.internalAU!.stopNote(noteNumber)
     }
 }
